Check mimetype and handle single upload

+++ src/index.js
@@ -36,21 +36,31 @@ export function docxtract(req: Request, res: Response): void {
   const uploads = [];
   const tmpdir = os.tmpdir();
   busboy.on('file', (fieldname, file, filename, encoding, mimetype) => {
-    const filepath = path.join(tmpdir, filename);
-    uploads.push({ fieldname, filepath, mimetype });
-    file.pipe(fs.createWriteStream(filepath));
+    if (
+      mimetype ===
+      'application/vnd.openxmlformats-officedocument.wordprocessingml.document'
+    ) {
+      const filepath = path.join(tmpdir, filename);
+      uploads.push({ fieldname, filepath });
+      file.pipe(fs.createWriteStream(filepath));
+    } else {
+      file.resume();
+    }
   });
   busboy.on('finish', async () => {
+    if (uploads.length === 0) {
+      return res.status(400).send({ error: 'Unknown mimetype' });
+    }
     const extractHandler = R.compose(
       R.composeP(extract, promisify(fs.readFile)),
       R.prop('filepath')
     );
-    const extracted = await Promise.all(R.map(extractHandler, uploads));
+    const extracted = await extractHandler(uploads[0]);
 
     const removeHandler = R.compose(promisify(fs.unlink), R.prop('filepath'));
     await Promise.all(R.map(removeHandler, uploads));
 
-    res.send(extracted);
+    return res.send(extracted);
   });
   if (req.rawBody) {
     busboy.end(req.rawBody);
